+++ services/modelserving/model_token_created.go
@@ -102,6 +102,33 @@ type TokenCreatedGetIdArgType = string
 type TokenCreatedGetIdRetType = string
 
 /*
+	types and functions for labels
+*/
+
+// isContainer
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type TokenCreatedGetLabelsAttributeType = *map[string]string
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type TokenCreatedGetLabelsArgType = map[string]string
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+type TokenCreatedGetLabelsRetType = map[string]string
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func getTokenCreatedGetLabelsAttributeTypeOk(arg TokenCreatedGetLabelsAttributeType) (ret TokenCreatedGetLabelsRetType, ok bool) {
+	if arg == nil {
+		return ret, false
+	}
+	return *arg, true
+}
+
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func setTokenCreatedGetLabelsAttributeType(arg *TokenCreatedGetLabelsAttributeType, val TokenCreatedGetLabelsRetType) {
+	*arg = &val
+}
+
+/*
 	types and functions for name
 */
 
@@ -340,6 +367,8 @@ type TokenCreated struct {
 	Description TokenCreatedGetDescriptionAttributeType `json:"description,omitempty"`
 	// REQUIRED
 	Id TokenCreatedGetIdAttributeType `json:"id" required:"true"`
+	// Object that represents the labels of an object. Regex for keys: `^(?=.{1,63}$)([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9]$`. Regex for values: `^(?=.{0,63}$)(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])*$`. Providing a `null` value for a key will remove that key. Send empty object {} to remove all labels. The `stackit` prefix is reserved and cannot be used for Keys.
+	Labels TokenCreatedGetLabelsAttributeType `json:"labels,omitempty"`
 	// REQUIRED
 	Name TokenCreatedGetNameAttributeType `json:"name" required:"true"`
 	// REQUIRED
@@ -445,6 +474,33 @@ func (o *TokenCreated) SetId(v TokenCreatedGetIdRetType) {
 	setTokenCreatedGetIdAttributeType(&o.Id, v)
 }
 
+// GetLabels returns the Labels field value if set, zero value otherwise.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *TokenCreated) GetLabels() (res TokenCreatedGetLabelsRetType) {
+	res, _ = o.GetLabelsOk()
+	return
+}
+
+// GetLabelsOk returns a tuple with the Labels field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *TokenCreated) GetLabelsOk() (ret TokenCreatedGetLabelsRetType, ok bool) {
+	return getTokenCreatedGetLabelsAttributeTypeOk(o.Labels)
+}
+
+// HasLabels returns a boolean if a field has been set.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *TokenCreated) HasLabels() bool {
+	_, ok := o.GetLabelsOk()
+	return ok
+}
+
+// SetLabels gets a reference to the given map[string]string and assigns it to the Labels field.
+// Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
+func (o *TokenCreated) SetLabels(v TokenCreatedGetLabelsRetType) {
+	setTokenCreatedGetLabelsAttributeType(&o.Labels, v)
+}
+
 // GetName returns the Name field value
 // Deprecated: Will be removed after 2026-09-30. Move to the packages generated for each available API version instead
 func (o *TokenCreated) GetName() (ret TokenCreatedGetNameRetType) {
@@ -536,6 +592,9 @@ func (o TokenCreated) ToMap() (map[string]interface{}, error) {
 	}
 	if val, ok := getTokenCreatedGetIdAttributeTypeOk(o.Id); ok {
 		toSerialize["Id"] = val
+	}
+	if val, ok := getTokenCreatedGetLabelsAttributeTypeOk(o.Labels); ok {
+		toSerialize["Labels"] = val
 	}
 	if val, ok := getTokenCreatedGetNameAttributeTypeOk(o.Name); ok {
 		toSerialize["Name"] = val